@@ -26,9 +26,9 @@
 @if (Context.User.Identity.IsAuthenticated)
 {
     <div class="text-danger">@($"{userId}({userDataInfos[2]})님의 권한은 {userRoleId}({userDataInfos[0]})입니다.")</div>
-    <a asp-controller="Membership" asp-action="LogOutAsync" class="btn btn-primary">멤버십 로그아웃</a>
+    <a asp-controller="Membership" asp-action="LogOut" class="btn btn-primary">멤버십 로그아웃</a>
 
-    <form method="post" asp-controller="Membership" asp-action="WithdrawnAsync" onsubmit="btnWithdrawn.disabled = true; return true;">
+    <form method="post" asp-controller="Membership" asp-action="Withdrawn" onsubmit="btnWithdrawn.disabled = true; return true;">
         <input type="hidden" name="UserId" value="@Context.User.Identity.Name" />
         <input type="password" name="Password" placeholder="Password" value="" autofocus/>
         <button type="submit" name="btnWithdrawn" class="btn btn-default">멤버십 사용자 탈퇴</button>
