@@ -15,7 +15,7 @@
 
 <div class="row">
     <div class="col-md-4">
-        <form method="post" asp-route-returnrul="@ViewData["ReturnUrl"]" asp-controller="Membership" asp-action="LoginAsync" onsubmit="btnLogin.disabled = true; return true;">
+        <form method="post" asp-route-returnrul="@ViewData["ReturnUrl"]" asp-controller="Membership" asp-action="Login" onsubmit="btnLogin.disabled = true; return true;">
             <div class="text-danger" asp-validation-summary="All"></div>
             <div class="form-group">
                 @Html.DisplayNameFor(m => m.UserId)
